Update tiptap to 1.30.0
Refactoring, move to dedicated files:
- TodoItem
- KeyboardSubmit (ctrl+enter)
