+++ client/components/widgets/richEditor/RichEditor.vue
@@ -29,8 +29,12 @@
 <script>
 import { Coeditions } from "/imports/api/coeditions/coeditions";
 
-import { Editor, EditorContent, Node, Extension } from "tiptap";
+import { Editor, EditorContent } from "tiptap";
+import KeyboardSubmit from "./extensions/KeyboardSubmit";
+import TodoItem from "./extensions/TodoItem";
+
 import {
+  Link,
   HardBreak,
   Blockquote,
   CodeBlock,
@@ -41,7 +45,6 @@ import {
   Bold,
   Code,
   Italic,
-  Link,
   Strike,
   Underline,
   History,
@@ -52,124 +55,6 @@ import {
   TableRow,
   Collaboration
 } from "tiptap-extensions";
-
-import {
-  sinkListItem,
-  splitToDefaultListItem,
-  liftListItem
-} from "tiptap-commands";
-
-/* eslint max-classes-per-file: off */
-/* eslint class-methods-use-this: off */
-class TodoItem extends Node {
-  get name() {
-    return "todo_item";
-  }
-
-  get defaultOptions() {
-    return {
-      nested: false
-    };
-  }
-
-  get view() {
-    return {
-      props: ["node", "updateAttrs", "view"],
-      methods: {
-        onChange() {
-          this.updateAttrs({
-            done: !this.node.attrs.done
-          });
-        }
-      },
-      template: `
-        <li :data-type="node.type.name" :data-done="node.attrs.done.toString()" data-drag-handle>
-          <span class="todo-checkbox" contenteditable="false" @click="onChange"></span>
-          <div class="todo-content" ref="content" :contenteditable="view.editable.toString()"></div>
-        </li>
-      `,
-      /*
-        The render function enables TodoItem to work in `runtimeonly` builds,
-        which is required for frameworks requiring strict CSP policies. For
-        example, doing this is required in Chrome Extensions. Having both
-        the template and the render function ensures there are no issues
-        converting the node to JSON and rendering the component.
-      */
-      render(h) {
-        return h(
-          "li",
-          {
-            attrs: {
-              "data-type": this.node.type.name,
-              "data-done": this.node.attrs.done.toString(),
-              "data-drag-handle": ""
-            }
-          },
-          [
-            h("span", {
-              class: "todo-checkbox",
-              attrs: {
-                contenteditable: false
-              },
-              on: {
-                click: this.onChange
-              }
-            }),
-            h("div", {
-              class: "todo-content",
-              attrs: {
-                contenteditable: this.view.editable.toString()
-              },
-              ref: "content"
-            })
-          ]
-        );
-      }
-    };
-  }
-
-  get schema() {
-    return {
-      attrs: {
-        done: {
-          default: false
-        }
-      },
-      draggable: true,
-      content: this.options.nested ? "(paragraph|todo_list)+" : "paragraph+",
-      toDOM: (node) => {
-        const { done } = node.attrs;
-
-        return [
-          "li",
-          {
-            "data-type": this.name,
-            "data-done": done.toString()
-          },
-          ["span", { class: "todo-checkbox", contenteditable: "false" }],
-          ["div", { class: "todo-content" }, 0]
-        ];
-      },
-      parseDOM: [
-        {
-          priority: 51,
-          tag: `[data-type="${this.name}"]`,
-          getAttrs: (dom) => ({
-            done: dom.getAttribute("data-done") === "true"
-          })
-        }
-      ]
-    };
-  }
-
-  keys({ type }) {
-    return {
-      Enter: splitToDefaultListItem(type),
-      Tab: this.options.nested ? sinkListItem(type) : () => {},
-      "Shift-Tab": liftListItem(type)
-    };
-  }
-}
 
 export default {
   components: {
@@ -274,8 +159,6 @@ export default {
     }
   },
   mounted() {
-    const vm = this;
-
     const extensions = [
       new HardBreak(),
       new Blockquote(),
@@ -301,15 +184,9 @@ export default {
       new TableHeader(),
       new TableCell(),
       new TableRow(),
-      new (class extends Extension {
-        keys() {
-          return {
-            "Ctrl-Enter": () => {
-              vm.submit();
-            }
-          };
-        }
-      })()
+      new KeyboardSubmit({
+        context: this
+      })
     ];
 
     if (this.collaboration) {
