+++ client/components/widgets/richEditor/extensions/KeyboardSubmit.js
@@ -1,0 +1,11 @@
+import { Extension } from "tiptap";
+
+export default class RichEditorKeyboardSubmit extends Extension {
+  keys() {
+    return {
+      "Ctrl-Enter": () => {
+        this.options.context.submit();
+      }
+    };
+  }
+}
