+++ package.json
@@ -38,8 +38,8 @@
     "showdown": "^1.9.1",
     "simpl-schema": "^1.7.0",
     "sortablejs": "^1.10.2",
-    "tiptap": "^1.27.1",
-    "tiptap-extensions": "^1.29.1",
+    "tiptap": "^1.30.0",
+    "tiptap-extensions": "^1.33.2",
     "v-click-outside": "^2.1.5",
     "vue": "^2.6.11",
     "vue-dragscroll": "^1.10.2",
